+++ prisma/migrations/20260405032619_npm_run_dev/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "SiteSettings" ADD COLUMN     "volunteerInterests" JSONB NOT NULL DEFAULT '["Fundraising","Mentoring","Event Support","Administrative","Technical"]';